init: new file init.c to have common initialisation
For all modes : fuzz, unit tests, real suricata, etc...
